fixed number issues

+++ Bot/cogs/team_generator.py
@@ -37,7 +37,7 @@ class UserDropdown(discord.ui.UserSelect):
     async def callback(self, interaction: discord.Interaction) -> Any:
         members: List[Union[discord.Member, discord.User]] = self.values
         random.shuffle(members)
-        for team_no in range(np.ceil(len(members) / self.team_size)):
+        for team_no in range(int(np.ceil(len(members) / self.team_size))):
             t1 = members[0:5]
             [members.remove(x) for x in members]  # pylint: disable=W0106
             random.shuffle(members)
